+++ week1/app/src/main/java/com/example/week1/data/FollowerData.kt
@@ -1,4 +1,4 @@
-package com.example.week1
+package com.example.week1.data
 
 import java.lang.invoke.MethodHandleInfo
