+++ week1/app/src/main/java/com/example/week1/data/RepoData.kt
@@ -1,4 +1,4 @@
-package com.example.week1
+package com.example.week1.data
 
 data class RepoData(
     val repo : String,
