+++ week1/app/src/main/java/com/example/week1/ui/EmptyFollowingFragment.kt
@@ -1,10 +1,11 @@
-package com.example.week1
+package com.example.week1.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.week1.R
 
 // TODO: Rename parameter arguments, choose names that match
 // the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
