+++ week1/app/src/main/java/com/example/week1/ui/FollowerFragment.kt
@@ -1,4 +1,4 @@
-package com.example.week1
+package com.example.week1.ui
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -7,6 +7,8 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.example.week1.ui.adapter.FollowerAdapter
+import com.example.week1.data.FollowerData
 import com.example.week1.databinding.FragmentFollowerBinding
