+++ week1/app/src/main/java/com/example/week1/ui/ProfileFragment.kt
@@ -1,17 +1,13 @@
-package com.example.week1
+package com.example.week1.ui
 
-import android.content.res.ColorStateList
-import android.graphics.Color
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.content.ContextCompat
-import androidx.fragment.app.FragmentManager
 import com.bumptech.glide.Glide
+import com.example.week1.R
 import com.example.week1.databinding.FragmentProfileBinding
-import com.example.week1.databinding.FragmentRepoBinding
 
 class ProfileFragment : Fragment() {
     private var position = FOLLOWER_POSITION
