+++ week1/app/src/main/java/com/example/week1/ui/SignInActivity.kt
@@ -1,4 +1,4 @@
-package com.example.week1
+package com.example.week1.ui
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
