+++ week1/app/src/main/java/com/example/week1/ui/adapter/HomeViewPagerAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.week1
+package com.example.week1.ui.adapter
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
