+++ week1/app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     android:id="@+id/frameLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfileFragment">
+    tools:context=".ui.ProfileFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:id="@+id/constraintLayout"
@@ -113,7 +113,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/container_home"
-        android:name="com.example.week1.FollowerFragment"
+        android:name="com.example.week1.ui.FollowerFragment"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
