+++ week1/app/src/main/res/layout/fragment_repo.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".RepoFragment">
+    tools:context=".ui.RepoFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_repo"
